EES-3782 Make Publication.IsLatestPublishedVersionOfRelease public
