+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/Extensions/ReleaseExtensions.cs
@@ -1,6 +1,5 @@
 ﻿#nullable enable
 using System;
-using System.Linq;
 using static GovUk.Education.ExploreEducationStatistics.Common.Model.FileType;
 using static GovUk.Education.ExploreEducationStatistics.Common.Services.FileStoragePathUtils;
 
@@ -35,20 +34,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Extensions
         /// <returns>True if the Release is the latest published version of a Release</returns>
         public static bool IsLatestPublishedVersionOfRelease(this Release release)
         {
-            if (release.Publication?.Releases == null || !release.Publication.Releases.Any())
-            {
-                throw new ArgumentException(
-                    "Release must be hydrated with Publications Releases to test the latest published version");
-            }
-
-            return
-                // Release itself must be live
-                release.Live
-                // It must also be the latest version unless the later version is a draft
-                && !release.Publication.Releases.Any(r =>
-                    r.Live
-                    && r.PreviousVersionId == release.Id
-                    && r.Id != release.Id);
+            return release.Publication.IsLatestPublishedVersionOfRelease(release);
         }
     }
 }
